+++ src/engine/cache_result.rs
@@ -2,12 +2,12 @@
 /// AssetCacheResult represents a cached assets
 /// physical position within the cache
 pub struct AssetCacheResult {
-    pub index: usize,
-    pub length: usize
+    pub index: u32,
+    pub length: u32
 }
 
 impl<'a> AssetCacheResult {
-    pub fn new(index: usize, length: usize) -> AssetCacheResult {
+    pub fn new(index: u32, length: u32) -> AssetCacheResult {
         AssetCacheResult {
             index: index,
             length: length
